move into util module
